remove dependency on 'notebook'
This is not actually required for using qsymm. Anyone installing
qsymm via pip is not going to have access to the notebooks anyway.
Anyone using qsymm by cloning the repository would be better served
by a conda 'environment.yml' file in the repository root (this could
also include e.g. plotting and Kwant).